removed some variables from context and refined some names
