+++ frontend/src/screens/dashboard/docsqa/main/context.tsx
@@ -6,20 +6,13 @@ import React, {
   ReactNode,
   useContext,
 } from 'react'
-import { SSE } from 'sse.js'
 
 import {
-  CollectionQueryDto,
   SourceDocs,
-  baseQAFoundryPath,
-  useCreateApplicationMutation,
   useGetAllEnabledChatModelsQuery,
   useGetCollectionNamesQuery,
   useGetOpenapiSpecsQuery,
 } from '@/stores/qafoundry'
-
-import { notifyError } from '@/utils/error'
-import notify from '@/components/base/molecules/Notify'
 
 import { DocsQAContextType, SelectedRetrieverType } from './types'
 
@@ -52,25 +45,20 @@ export const DocsQAProvider: React.FC<DocsQAProviderProps> = ({ children }) => {
   const [selectedRetriever, setSelectedRetriever] = useState<
     SelectedRetrieverType | undefined
   >()
-  const [prompt, setPrompt] = useState('')
-  const [isRunningPrompt, setIsRunningPrompt] = useState(false)
-  const [answer, setAnswer] = useState('')
-  const [sourceDocs, setSourceDocs] = useState<SourceDocs[]>([])
-  const [errorMessage, setErrorMessage] = useState(false)
-  const [modelConfig, setModelConfig] = useState(defaultModelConfig)
-  const [retrieverConfig, setRetrieverConfig] = useState(defaultRetrieverConfig)
-  const [promptTemplate, setPromptTemplate] = useState(defaultPrompt)
 
   const [isInternetSearchEnabled, setIsInternetSearchEnabled] = useState(false)
-  const [isCreateApplicationModalOpen, setIsCreateApplicationModalOpen] =
-    useState(false)
+  const [retrieverConfig, setRetrieverConfig] = useState(defaultRetrieverConfig)
+  const [modelConfig, setModelConfig] = useState(defaultModelConfig)
+  const [promptTemplate, setPromptTemplate] = useState(defaultPrompt)
+  const [sourceDocs, setSourceDocs] = useState<SourceDocs[]>([])
+  const [errorMessage, setErrorMessage] = useState(false)
+  const [answer, setAnswer] = useState('')
+  const [prompt, setPrompt] = useState('')
 
   const { data: collections, isLoading: isCollectionsLoading } =
     useGetCollectionNamesQuery()
   const { data: allEnabledModels } = useGetAllEnabledChatModelsQuery()
   const { data: openapiSpecs } = useGetOpenapiSpecsQuery()
-  const [createApplication, { isLoading: isCreateApplicationLoading }] =
-    useCreateApplicationMutation()
 
   const allQueryControllers = useMemo(() => {
     if (!openapiSpecs?.paths) return []
@@ -97,128 +85,6 @@ export const DocsQAProvider: React.FC<DocsQAProviderProps> = ({ children }) => {
       promptTemplate: value.value.prompt_template ?? defaultPrompt,
     }))
   }, [selectedQueryController, openapiSpecs])
-
-  const handlePromptSubmit = async () => {
-    setIsRunningPrompt(true)
-    setAnswer('')
-    setSourceDocs([])
-    setErrorMessage(false)
-    try {
-      const selectedModel = allEnabledModels.find(
-        (model: any) => model.name == selectedQueryModel,
-      )
-      if (!selectedModel) {
-        throw new Error('Model not found')
-      }
-      try {
-        JSON.parse(modelConfig)
-      } catch (err: any) {
-        throw new Error('Invalid Model Configuration')
-      }
-      try {
-        JSON.parse(retrieverConfig)
-      } catch (err: any) {
-        throw new Error('Invalid Retriever Configuration')
-      }
-
-      const params: CollectionQueryDto = Object.assign(
-        {
-          collection_name: selectedCollection,
-          query: prompt,
-          model_configuration: {
-            name: selectedModel.name,
-            provider: selectedModel.provider,
-            ...JSON.parse(modelConfig),
-          },
-          retriever_name: selectedRetriever?.name ?? '',
-          retriever_config: JSON.parse(retrieverConfig),
-          prompt_template: promptTemplate,
-          internet_search_enabled: isInternetSearchEnabled,
-        },
-        {},
-      )
-
-      const sseRequest = new SSE(
-        `${baseQAFoundryPath}/retrievers/${selectedQueryController}/answer`,
-        {
-          payload: JSON.stringify({
-            ...params,
-            stream: true,
-          }),
-          headers: {
-            'Content-Type': 'application/json',
-          },
-        },
-      )
-
-      sseRequest.addEventListener('data', (event: any) => {
-        try {
-          const parsed = JSON.parse(event.data)
-          if (parsed?.type === 'answer') {
-            setAnswer((prevAnswer) => prevAnswer + parsed.content)
-            setIsRunningPrompt(false)
-          } else if (parsed?.type === 'docs') {
-            setSourceDocs((prevDocs) => [...prevDocs, ...parsed.content])
-          }
-        } catch (err: any) {
-          throw new Error('An error occurred while processing the response.')
-        }
-      })
-
-      sseRequest.addEventListener('end', (event: any) => {
-        sseRequest.close()
-      })
-
-      sseRequest.addEventListener('error', (event: any) => {
-        sseRequest.close()
-        setPrompt('')
-        setIsRunningPrompt(false)
-        const message = JSON.parse(event.data).detail[0].msg
-        notifyError('Failed to retrieve answer', { message })
-      })
-    } catch (err: any) {
-      setPrompt('')
-      setIsRunningPrompt(false)
-      notifyError('Failed to retrieve answer', err)
-    }
-  }
-
-  const createChatApplication = async (
-    applicationName: string,
-    questions: [],
-    setApplicationName: (name: string) => void,
-  ) => {
-    if (!applicationName) {
-      return notify('error', 'Application name is required')
-    }
-    const selectedModel = allEnabledModels.find(
-      (model: any) => model.name == selectedQueryModel,
-    )
-
-    try {
-      await createApplication({
-        name: `${applicationName}-rag-app`,
-        config: {
-          collection_name: selectedCollection,
-          model_configuration: {
-            name: selectedModel.name,
-            provider: selectedModel.provider,
-            ...JSON.parse(modelConfig),
-          },
-          retriever_name: selectedRetriever?.name ?? '',
-          retriever_config: JSON.parse(retrieverConfig),
-          prompt_template: promptTemplate,
-          query_controller: selectedQueryController,
-        },
-        questions,
-      }).unwrap()
-      setApplicationName('')
-      setIsCreateApplicationModalOpen(false)
-      notify('success', 'Application created successfully')
-    } catch (err: any) {
-      notify('error', 'Failed to create application', err?.data?.detail)
-    }
-  }
 
   const resetQA = () => {
     setAnswer('')
@@ -251,42 +117,35 @@ export const DocsQAProvider: React.FC<DocsQAProviderProps> = ({ children }) => {
 
   const value = {
     selectedQueryModel,
-    setSelectedQueryModel,
     selectedCollection,
-    setSelectedCollection,
     selectedQueryController,
-    setSelectedQueryController,
     selectedRetriever,
-    setSelectedRetriever,
     prompt,
-    setPrompt,
-    isRunningPrompt,
-    setIsRunningPrompt,
     answer,
-    setAnswer,
     sourceDocs,
-    setSourceDocs,
     errorMessage,
-    setErrorMessage,
     modelConfig,
-    setModelConfig,
     retrieverConfig,
-    setRetrieverConfig,
     promptTemplate,
-    setPromptTemplate,
     isInternetSearchEnabled,
-    setIsInternetSearchEnabled,
-    isCreateApplicationModalOpen,
-    setIsCreateApplicationModalOpen,
     collections,
     isCollectionsLoading,
     allEnabledModels,
     allQueryControllers,
     allRetrieverOptions,
-    handlePromptSubmit,
+    setSelectedQueryModel,
+    setSelectedCollection,
+    setSelectedQueryController,
+    setSelectedRetriever,
+    setSourceDocs,
+    setErrorMessage,
+    setModelConfig,
+    setRetrieverConfig,
+    setPromptTemplate,
+    setIsInternetSearchEnabled,
     resetQA,
-    isCreateApplicationLoading,
-    createChatApplication,
+    setPrompt,
+    setAnswer,
   }
 
   return (
